requirements.txt
alembic==1.0.10
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap4==4.0.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.18.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.9
MarkupSafe==1.1.1
numpy==1.16.3
pandas==0.24.2
PyExifTool==0.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.1
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.3
visitor==0.1.3
Werkzeug==0.15.2
WTForms==2.2.1
